feat(planner, executor): Support chain and set result in batch
In batch-processing, chain functions will not callBatch to planner
immediately, it will be grouped according to next fucntion name and
appid and stored in a queue, which will be called after function is
finish.

If a executor's request has more than one message, all the messages set
results to PlannerServer in one request.